refactor: move parseAlternateRegistries function to CargoTomlParser
- Relocated the `parseAlternateRegistries` function from `CargoTomlListener` to `CargoTomlParser` for better organization and separation of concerns.
- Updated imports in `CargoTomlListener` to reflect the new location of the parsing function.
